renamed styles file
